Examples and core updates
Some new examples added: using a main container element with vizi,
using child elements of the main scrolling element, visibility
checking. Also fixed some issues in core with events not firing if the
browser was scrolled to the bottom. Added in an option to check if the
whole element was visible or just partially.